pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fingerprints-armor"
version = "1.0.0"
description = "A package implementing the NEURAL FINGERPRINTS FOR ADVERSARIAL ATTACK DETECTION"
authors = [
    { name = "Haim D. Fisher", email = "haimdfisher@gmail.com" },
    { name = "Yehezkel S. Resheff" },
    { name = "Moni Shahar" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.1.2",
    "torch>=2.5.0",
    "cleverhans>=4.0.0",
    "Pillow>=11.0.0",
    "timm>=1.0.11",
    "pandas>=2.2.3",
    "scikit-learn>=1.5.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=4.0.0"
]

[project.urls]
Homepage = "https://github.com/HaimFisher/fingerprints-armor"
Documentation = "https://github.com/HaimFisher/fingerprints-armor/blob/main/README.md"

[tool.setuptools.packages.find]
where = ["src"]
